renaming the CL in 2D
