+++ applications/ConstitutiveLawsApplication/constitutive_laws_application.cpp
@@ -72,7 +72,7 @@ void KratosConstitutiveLawsApplication::Register()
     KRATOS_REGISTER_CONSTITUTIVE_LAW("SerialParallelRuleOfMixturesLaw", mSerialParallelRuleOfMixturesLaw);
 
     // Anisotropic law
-    KRATOS_REGISTER_CONSTITUTIVE_LAW("GenericAnisotropic2DLaw", mGenericAnisotropic2DLaw);
+    KRATOS_REGISTER_CONSTITUTIVE_LAW("GenericAnisotropicPlaneStrain2DLaw", mGenericAnisotropicPlaneStrain2DLaw);
     KRATOS_REGISTER_CONSTITUTIVE_LAW("GenericAnisotropic3DLaw", mGenericAnisotropic3DLaw);
 
     /// Plasticity
