+++ applications/ConstitutiveLawsApplication/constitutive_laws_application.h
@@ -66,7 +66,7 @@
 #include "custom_constitutive/small_strains/plastic_damage/generic_small_strain_plastic_damage_model.h"
 #include "custom_constitutive/small_strains/damage/generic_small_strain_orthotropic_damage.h"
 #include "custom_constitutive/composites/serial_parallel_rule_of_mixtures_law.h"
-#include "custom_constitutive/small_strains/anisotropy_orthotropy/generic_anisotropic_3d_law.h"
+#include "custom_constitutive/small_strains/anisotropy_orthotropy/generic_anisotropic_law.h"
 #include "custom_constitutive/small_strains/linear/multi_linear_elastic_1d_law.h"
 #include "custom_constitutive/small_strains/linear/multi_linear_isotropic_plane_stress_2d.h"
 #include "custom_constitutive/small_strains/damage/generic_small_strain_isotropic_damage_plane_stress.h"
@@ -623,7 +623,7 @@ private:
 	const ParallelRuleOfMixturesLaw<2> mParallelRuleOfMixturesLaw2D;
 
     // Anisotropic law
-    const GenericAnisotropicLaw<2> mGenericAnisotropic2DLaw;
+    const GenericAnisotropicLaw<2> mGenericAnisotropicPlaneStrain2DLaw;
     const GenericAnisotropicLaw<3> mGenericAnisotropic3DLaw;
 
     const AssociativePlasticDamageModel <VonMisesYieldSurface<VonMisesPlasticPotential<6>>> mAssociativePlasticDamageModel3DVonMises;
